+++ app/pages/admin/nextauth/githubsSignin.ts
@@ -1,5 +1,5 @@
 "use server";
-import { signIn } from "@/auth";
+import { signIn } from "@/auth-old";
 
 export default async function action() {
   return await signIn("github");
