+++ app/pages/contact/page.tsx
@@ -1,20 +1,22 @@
-import { signIn, signOut } from "@/auth";
-
-export default function Tech() {
-  const github_action = async () => {
-    "use server";
-    await signIn("github", {
-      callbackUrl: "https://v9qy5n-3000.csb.app/api/auth",
-    });
-  };
+import { signInS } from "./signin";
+/*        <form action={signInS ("github")}>
+          <button type="sumit">GitHub</button>
+        </form>
+        */
+export default async  function Tech() {
   return (
     <div className="">
       <div>
         <p className="text-4xl m-2">
           contact GuangZhou WanWei email : joe@wwin.cn
         </p>
-        <form action={github_action}>
-          <button type="sumit">GitHub</button>
+      ?</div>
+      <div>
+        <p className="text-4xl m-2">
+          Google
+        </p>
+        <form action={signInS}>
+          <button type="sumit">Google</button>
         </form>
       </div>
     </div>
