--- middleware.ts
@@ -1,18 +1,0 @@
-import NextAuth from "next-auth";
-import { authConfig } from "./auth.config";
-
-export default NextAuth(authConfig).auth; // or auth in auth.ts ?
-
-export const config = {
-  // https://nextjs.org/docs/app/building-your-application/routing/middleware#matcher
-  matcher: ["/((?!api|_next/static|_next/image|.*\\.png$).*)"],
-};
-
-/*import NextAuth from "next-auth";
-import { authConfig } from "./auth.config";
-
-export default NextAuth(authConfig).auth;
-
-export const config = {
-  matcher: ["/((?!api|_next/static|_next/image|.*\\.png$).*)"],
-};*/
