+++ package.json
@@ -26,7 +26,7 @@
     "fs": "^0.0.1-security",
     "multer": "^1.4.5-lts.1",
     "next": "^14.0.2",
-    "next-auth": "^5.0.0-beta.15",
+    "next-auth": "^5.0.0-beta.18",
     "postcss": "8.4.31",
     "react": "18.2.0",
     "react-dom": "18.2.0",
